Renamed files to not small character
