+++ src/software/hrvo/BUILD
@@ -1,6 +1,6 @@
 package(default_visibility = [
-    "//software/HRVO:__subpackages__",
     "//software/ai:__subpackages__",
+    "//software/hrvo:__subpackages__",
 ])
 
 cc_test(
@@ -17,118 +17,18 @@ cc_test(
 cc_library(
     name = "hrvo",
     srcs = [
-        "Agent.cpp",
-        "Goal.cpp",
-        "KdTree.cpp",
-        "Simulator.cpp",
-        "Vector2.cpp",
+        "agent.cpp",
+        "goal.cpp",
+        "kd_tree.cpp",
+        "simulator.cpp",
+        "vector2.cpp",
     ],
     hdrs = [
-        "Agent.h",
-        "Goal.h",
-        "KdTree.h",
-        "Simulator.h",
-        "Vector2.h",
-        #        "HRVO.h",
+        "agent.h",
+        "goal.h",
+        "kd_tree.h",
+        "simulator.h",
+        "vector2.h",
     ],
-    #    copts = [
-    #        "-fvisibility-inlines-hidden",
-    #        "-fvisibility=hidden",
-    #    ],
-    #    includes = ["."],
     visibility = ["//visibility:public"],
 )
-
-#filegroup(
-#    name = "hdrs",
-#    srcs = [
-#        # "Export.h",
-#        "HRVO.h",
-#        "Simulator.h",
-#        "Vector2.h",
-#    ],
-#    visibility = ["//visibility:private"],
-#)
-
-#cc_library(
-#    name = "hrvo",
-#    srcs = [
-#        "Agent.cpp",
-#        "Agent.h",
-#        "Goal.cpp",
-#        "Goal.h",
-#        "KdTree.cpp",
-#        "KdTree.h",
-#        "Simulator.cpp",
-#        "Vector2.cpp",
-#    ],
-#    hdrs = [":hdrs"],
-#    copts = [
-#        "-fvisibility-inlines-hidden",
-#        "-fvisibility=hidden",
-#    ],
-#    includes = ["."],
-#    visibility = ["//visibility:public"],
-#)
-
-#############################################
-# cc_library(
-#     name = "Agent",
-#     srcs = ["Agent.cpp"],
-#     hdrs = ["Agent.h"],
-#     deps = [
-#         ":Vector2",
-#         ":Goal",
-#         ":Simulator",
-#     ],
-# )
-
-# cc_library(
-#     name = "Definition",
-#     hdrs = ["Definition.h"],
-# )
-
-# cc_library(
-#     name = "Goal",
-#     srcs = ["Goal.cpp"],
-#     hdrs = ["Goal.h"],
-#     deps = [":Vector2"],
-# )
-
-# cc_library(
-#     name = "KdTree",
-#     srcs = ["KdTree.cpp"],
-#     hdrs = ["KdTree.h"],
-#     deps = [
-#         ":Vector2",
-#         ":Agent",
-#     ],
-# )
-
-# cc_library(
-#     name = "Simulator",
-#     srcs = ["Simulator.cpp"],
-#     hdrs = ["Simulator.h"],
-#     deps = [
-#         ":Vector2",
-#         ":Goal",
-#         ":Agent",
-#         ":KdTree",
-#     ],
-# )
-
-# cc_library(
-#     name = "Vector2",
-#     srcs = ["Vector2.cpp"],
-#     hdrs = ["Vector2.h"],
-#     # deps = [":Vector2"],
-# )
-
-# cc_library(
-#     name = "HRVO",
-#     hdrs = ["HRVO.h"],
-#     deps = [
-#         ":Simulator",
-#         ":Vector2",
-#     ],
-# )
