--- src/software/hrvo/HRVO.h
@@ -1,4 +1,0 @@
-#pragma once
-
-#include "Simulator.h"
-#include "Vector2.h"
