+++ src/software/hrvo/goal.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <vector>
 
-#include "Vector2.h"
+#include "vector2.h"
 
 class Simulator;
