+++ src/software/hrvo/kd_tree.h
@@ -3,7 +3,7 @@
 #include <cstddef>
 #include <vector>
 
-#include "Vector2.h"
+#include "vector2.h"
 
 class Agent;
 class Simulator;
