+++ src/software/hrvo/vector2.cpp
@@ -1,4 +1,4 @@
-#include "Vector2.h"
+#include "vector2.h"
 
 #include <ostream>
